Fix issue with currency translation
Issue #4
